requirements.txt
paho-mqtt
netifaces
click
